book-to-game/pyproject.toml
[project]
name = "book-to-game"
version = "0.1.0"
description = "Transform books into interactive games and AR experiences for Unity"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.84.0",
    "python-dotenv>=1.1.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
